Report empty indented code lines
The VS editor will now see an empty artifact when
the user types four spaces (a destructive change)
and expand the artifact as the user adds content.

TryConsumeContentLine() is now properly aware of
empty lines that are followed by an EOF or block
end, as opposed to lines which end of the block.
This eliminates the earlier ugly hack for fenced
code blocks with trailing blank lines at EOF; it
will now consume them as an empty line with '\0'

This also fixes lines with exactly one character
(which used to be swallowed because NextChar was
a newline).