Try to expose IWebEditorFactory for Markdown
On the latest VS2013 build, opening an MD file
doesn't use the HtmlEditorFactory to create an
editor. This makes the components of the HTMLX
editor initialize in the wrong order, breaking
IntelliSense and raising other issues too.

I tried to export an IWebEditorFactory, but
this comes from COM rather than MEF, and it
didn't help.
I think it needs to register .MD files with
the HTML editor's GUID?