Better capturing of region name in JavaScript regions
This is a fix to a small issue where the region name wouldn't be picked up when the region starts with // #region Name

When there is a space between // and #region, the region collapses (since startHide = "// #region"), however the regex that extracts the name does not allow the space, so when collapsed the region would show "#region" instead of the name. This new regex works with both //#region and // #region.