Setting to skip relative paths fix on css minify
Sharepoint has virtual "/Layout" folder, so it makes sense to keep there
resources, which are directly "deployed" to the server.
Outside this folder can be development going on, e.g. css-files, which
eventually need to be compiled, minified and so on.
Having enabled css relative paths overwriting on minifying by default
without a setting to turn it off interfere with this folder structure.

Say we have
/Layout/Project1/Styles/fonts/font-awesome.woff
/Layout/Project1/Styles/css/plugins.css.bundle

now, bundle, references file:
/Styles/css/font-awesome.css

now replacing the "../font.woff" to "../../../../font.woff".

The commit is rather simple and merely adds a Misc option to css  bundle
compile check inside
WriteBundleFile,
controlling thus
CssUrlNormalizer.NormalizeUrls

Feel free to commit, or point out why that's unwise, how could we make
this better.