Added a mechanism for customizing the namespace (i.e. module name) of TS interfaces generated from C# classes.
By default, the name of the module for generated interfaces is "server" (to provide backwards compatibility).
To specify a different name, one must apply an attribute named "TypeScriptModule" (name is case-insensitive) with a single constructor parameter of type string. That parameter will be used as the module name for generated interfaces.