SourceMap: Independent of settings.
The aim is to make source-map generation
independent of WE's "generate .map" settings
for LESS and SCSS..

It will compile just once on document open (which
it already does); if the map generation is set,
read from the generated map file. Otherwise,
in compiler parameter we set path to temp file,
read all text and delete the file.