Code: Reverting #994.
#994 is duplicate of #844 and its wrong.
This PR reverts it back.

We are creating source map for both LESS and SASS
regardless of settings. We are using these
interim maps for editor enhancements.

We extract out Base-64 VLQ values, decode it and
use that info for features which require
source-to-source mapping context (see #787).

After parsing map files & extracting the desired
info, the map-file is deleted; if generate-map
option is disabled.