Compiler: Test MinifyInPlace for source-map. Fixes #1070.
Story:

In VS compiler settings, MinifyInPlace is false
by default. If it is set to true and the target's
GenerateSourceMap setting is also true, the
compiler and its minifying listeners should skip
generating the source-map file (for min file).

In WE's world, mappings in .min.map corresponds
to the immediate source file; that is the
unobtrusive version of the same and not the
pre-processed script (LESS/SCSS/CoffeeScript/SJS)

Also, the only context resolver for pre-processed
script to the source-map is the compiler itself.
Since some compilers don't offer minified output
with source-map directly, it would be disparity
between behaviors, if we go that route. This
would also be rendered as disparity because of
the reasons mentioned in previous paragraph:
(what is .min.map in Web Essentials' jargon).

Therefore, this PR prevents the compiler and
minifier to avoid generating source-map, if
MinifyInPlace is enabled. This will also avoid
the confusions like #1070.