Bundle: Observe renaming and creation.
Recipe Creation:
When a recipe resource is created during the
project lifetime, FileSystemWatcher will start
monitoring the file. It caters the scenario when
we move a bundle file from one directory to
another within the projects under the current
solution.

Recipe Renaming:
When a recipe file is renamed, it detaches the
existing FSWs from the recipe file and all its
constituents and reattaches the new ones.

NOTE: this does not respond to the
constituents files' changes, because we do not
want to change the contents of the recipe files
programmatically.