Fix Drag&Drop images in Markdown documents(issue #1745)
- Renamed the HtmlImageDropHandlerProvider to MarkdownImageDropHandlerProvider
and set the ContentType to MarkdownContentType since now the functionality for Html
documents is provided out of the box by Visual Studio.

- Added [Order(Before = "HtmlViewFileDropHandlerProvider")] on the MarkdownImageDropHandlerProvider
to ensure that the provider is inserted before all the default VS implementations.
The HtmlViewFileDropHandlerProvider was added by Visual studio Update 4.
This made placing the [Order(Before = "DefaultFileDropHandler")] not enough to define a strict order
for the drop handler providers registered to handle the "CF_VSSTGPROJECTITEMS" drop format.
Depending on the order in witch the exported IDropHandlerProvider were discovered by MEF, the
HtmlViewFileDropHandlerProvider could end up before the MarkdownImageDropHandlerProvider.

- Moved code to handle image link generation in a common method (see ImageExtensions.InsertLinkToImageFile).
This commit include also some refactoring that make the behavior of Drag&Drop and
Copy&Paste images the same.
The url inserted for both cases will be always relative to the current document.
Before, when an image was dropped into a markdown document the generated url was
relative to the project folder that contained the markdown document.