Fixed encoding issue for non ANSI markdown files
see Issue #1762(https://github.com/madskristensen/WebEssentials2013/issues/1762)

When a file is created without the BOM FileHelpers.ReadAllTextRetry assume that the encoding is UTF8 and the returned string can contain gibberish.
The code now will make use of the encoding that visual studio has detected by reading it from the ITextDocument.Encoding property.
Note that this will not fix the issue completely (static compilation can still incur in the same problem), but  it should make it very rare, since the user will correct the issue while editing the file and after the file is saved with the BOM, File.ReadAllText will detect the right encoding to use.